Fixed an issue with Forge building knowledge during the Mojang screen
The issue was due to Forge loading the sound system before mods were loaded, which does not happen with ModLoader. Then, when MAtmos loads expansion files, it directly tries to turn them on during the Mojang screen (main thread) and does it sucessfully, causing the knowledges to be built in the main thread.
This was fixed by threading the turn on signal when all mods are loaded.